LayoutTests/http/tests/security/contentSecurityPolicy/audioworklet-script-src-allowed-expected.txt

CONSOLE MESSAGE: PASS: Audio worklet load was allowed by CSP
